refactor(path): convert path for types to relative
